Move memory related to mem_allocator
