+++ src/common/base/src/mem_allocator/je_allocator.rs
@@ -16,8 +16,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-pub static ALLOC: JEAllocator = JEAllocator;
-
 pub use platform::*;
 pub use tikv_jemalloc_sys;
