+++ src/common/base/tests/it/mem_allocator/mod.rs
@@ -12,5 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod runtime;
-mod runtime_tracker;
+mod allocator;
