+++ src/common/base/tests/it/runtimev2/runtime_tracker.rs
@@ -17,7 +17,7 @@ use std::alloc::Layout;
 use std::ptr::null_mut;
 
 use common_base::mem_allocator::JEAllocator;
-use common_base::runtime::ProxyAllocator;
+use common_base::mem_allocator::ProxyAllocator;
 use common_base::runtime::RuntimeTracker;
 
 static ALLOC: ProxyAllocator<JEAllocator> = ProxyAllocator::new(JEAllocator);
